requirements.txt
fastapi==0.110.0
python-dotenv==1.0.1
